moved course search function to the helper file, added sorting when searching
